+++ Cache_Set.cpp
@@ -6,7 +6,7 @@ ReplaceManager::ReplaceManager( uint32_t way, uint32_t rp ) :
         m_ReplacePolicy( rp ) {
     m_Record = new std::vector < uint8_t >() ;
     if ( m_ReplacePolicy == LRU )
-        for ( int i = 0; i < way; i++ )
+        for ( uint8_t i = 0; i < way; i++ )
             m_Record->push_back( i ) ;
     // TODO  else if ( m_RP == ROUND_ROBIN ) ;
 }  // ReplaceManager::ReplaceManager()
@@ -20,7 +20,7 @@ uint8_t ReplaceManager::GetReplaceIndex() {
 
 void ReplaceManager::UpdateRecord( uint8_t index, bool Endposition ) {
     if ( m_ReplacePolicy == LRU ) {
-        for ( int i = 0; i < m_Record->size(); i++ ) {
+        for ( uint8_t i = 0; i < m_Record->size(); i++ ) {
             if ( ( *m_Record )[ i ] == index ) {
                 m_Record->erase( m_Record->begin() + i ) ;
                 if ( Endposition )
@@ -39,7 +39,7 @@ Cache_Set::Cache_Set( uint32_t blocksize, uint32_t associativity, uint32_t repla
         m_BlockSize( blocksize ), m_Associativity( associativity ), m_WritePolicy( writePolicy ), m_RetentionTime( 0 ), m_ReadLatency( ReadLatency ), m_WriteLatency( WriteLatency ) {
     m_RP_Manager = new ReplaceManager( associativity, REPLACEPOLICY::LRU ) ;
     m_Way = new Way[ m_Associativity ] ;
-    for ( int i = 0; i < m_Associativity; i++ ) {
+    for ( uint8_t i = 0; i < m_Associativity; i++ ) {
         m_Way[ i ].mTag = -1 ;
         m_Way[ i ].Valid = false ;
         m_Way[ i ].Dirty = false ;
@@ -58,7 +58,7 @@ Cache_Set::Cache_Set( uint32_t blocksize, uint32_t associativity, uint32_t repla
         m_BlockSize( blocksize ), m_Associativity( associativity ), m_WritePolicy( writePolicy ), m_RetentionTime( retentionTime ), m_ReadLatency( ReadLatency ), m_WriteLatency( WriteLatency ) {
     m_RP_Manager = new ReplaceManager( associativity, REPLACEPOLICY::LRU ) ;
     m_Way = new Way[ m_Associativity ] ;
-    for ( int i = 0; i < m_Associativity; i++ ) {
+    for ( uint8_t i = 0; i < m_Associativity; i++ ) {
         m_Way[ i ].mTag = -1 ;
         m_Way[ i ].Valid = false ;
         m_Way[ i ].Dirty = false ;
@@ -91,7 +91,7 @@ void Cache_Set::AllocateData( Byte * in, uint64_t tag, uint32_t way_index, uint3
 }  // Cache_Set::WriteLine()
 
 uint32_t Cache_Set::FindTagInWay( uint64_t tag ) {
-    for ( int i = 0; i < m_Associativity; i++ ) {
+    for ( uint8_t i = 0; i < m_Associativity; i++ ) {
         if ( m_Way[ i ].mTag == tag && m_Way[ i ].Valid )
             return i ;
     }  // for
